+++ packages/typir/src/kinds/function/function-kind.ts
@@ -71,13 +71,13 @@ export interface InferFunctionCall<LanguageType = unknown, T extends LanguageTyp
      *
      * This property is specific for this function type.
      * If this function type is not overloaded, this property switches this validation off and on for the calls of this function,
-     * i.e. creates validation hints for all calls with mismatching argument types.
+     * i.e. creates validation issues for all calls with mismatching argument types.
      *
      * If this function type is overloaded, different values for this property for different overloaded functions might be specified:
-     * If the property is switched off for all overloads, no validation hints will be created.
-     * If the property is switched on for at least one overload, validation hints for will be shown for all calls (when none of the signatures match),
+     * If the property is switched off for all overloads, no validation issues will be created.
+     * If the property is switched on for at least one overload, validation issues for will be shown for all calls (when none of the signatures match),
      * since it is unclear, which of the overloads is the desired one!
-     * But the shown validation hint/message will not report about signatures for which this validation property is switched off.
+     * But the shown validation issue/message will not report about signatures for which this validation property is switched off.
      * While different values for this property for different overloads are possible in theory with the defined behaviour,
      * in practise this seems to be rarely useful.
      */