+++ packages/typir/test/kinds/class/class.test.ts
@@ -8,7 +8,7 @@ import { describe, test } from 'vitest';
 import { isClassType } from '../../../src/kinds/class/class-type.js';
 import { isPrimitiveType } from '../../../src/kinds/primitive/primitive-type.js';
 import { BooleanLiteral, ClassConstructorCall, ClassFieldAccess, IntegerLiteral, Variable } from '../../../src/test/predefined-language-nodes.js';
-import { createTypirServicesForTesting, expectToBeType, expectTypirTypes, expectValidationHints } from '../../../src/utils/test-utils.js';
+import { createTypirServicesForTesting, expectToBeType, expectTypirTypes, expectValidationIssues } from '../../../src/utils/test-utils.js';
 import { assertType } from '../../../src/utils/utils.js';
 
 describe('Tests some details for class types', () => {
@@ -45,9 +45,9 @@ describe('Tests some details for class types', () => {
             // infer the type when accessing fields
             .inferenceRuleForFieldAccess({
                 filter: node => node instanceof ClassFieldAccess,
-                matching: node => {
-                    const varType = typir.Inference.inferType(node.classVariable); // TODO review: doing type inference on your own here feels a bit strange
-                    return isClassType(varType) && varType.getName() === 'MyClass1';
+                matching: (node, classType) => {
+                    const variableType = typir.Inference.inferType(node.classVariable); // TODO review: doing type inference on your own here feels a bit strange
+                    return variableType === classType;
                 },
                 field: node => node.fieldName,
                 // a useless validation just for testing
@@ -64,17 +64,17 @@ describe('Tests some details for class types', () => {
 
         // var1 := new MyClass1();
         const varClass = new Variable('var1', new ClassConstructorCall('MyClass1'));
-        expectValidationHints(typir, varClass.initialValue, ["Called constructor for 'MyClass1'."]);
+        expectValidationIssues(typir, varClass.initialValue, ["Called constructor for 'MyClass1'."]);
 
         // var2 := var1.fieldInteger;
         const varFieldIntegerValue = new Variable('var2', new ClassFieldAccess(varClass, 'fieldInteger'));
         expectToBeType(typir.Inference.inferType(varFieldIntegerValue), isPrimitiveType, type => type.getName() === 'integer');
-        expectValidationHints(typir, varFieldIntegerValue.initialValue, []);
+        expectValidationIssues(typir, varFieldIntegerValue.initialValue, []);
 
         // var3 := var1.fieldBoolean;
         const varFieldBooleanValue = new Variable('var3', new ClassFieldAccess(varClass, 'fieldBoolean'));
         expectToBeType(typir.Inference.inferType(varFieldBooleanValue), isPrimitiveType, type => type.getName() === 'boolean');
-        expectValidationHints(typir, varFieldBooleanValue.initialValue, ["Validated access of 'fieldBoolean' of the variable 'var1'."]);
+        expectValidationIssues(typir, varFieldBooleanValue.initialValue, ["Validated access of 'fieldBoolean' of the variable 'var1'."]);
     });
 
 });